Improve cross-referencing macro expansion in preprocessor
A user reported a bug where a macro that references another macro
doesn't properly expand, which was due to the way I was doing recursive
expansion on macros. This addresses the case.

It also exposes there is another issue with the actual C reference for
macro expansion - where the preprocessor *still* does the wrong thing
with some weird macro edge cases. See `xtest`. That bug is for another
day!

Addresses #42